데이터 타입 맞추기 & Order Mock API (#8)

+++ src/main/kotlin/kr/hhplus/be/server/controller/coupon/dto/request/CouponIssueRequest.kt
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.*
 
 data class CouponIssueRequest(
     @field:NotNull(message = "계정 ID는 필수입니다")
-    @field:Min(value = 1, message = "계정 ID는 1 이상이어야 합니다")
-    val accountId: Int
+    @field:Min(value = 1L, message = "계정 ID는 1 이상이어야 합니다")
+    val accountId: Long
 )
