+++ src/main/kotlin/kr/hhplus/be/server/controller/product/dto/response/ProductResponse.kt
@@ -1,30 +1,29 @@
 package kr.hhplus.be.server.controller.product.dto.response
 
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class ProductResponse(
-    val productId: Int,
+    val productId: Long,
     val name: String,
     val description: String,
-    val price: BigDecimal,
+    val price: Double,
     val options: List<ProductOptionResponse>,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime
 )
 
 data class ProductOptionResponse(
-    val optionId: Int,
+    val optionId: Long,
     val name: String,
-    val additionalPrice: BigDecimal,
+    val additionalPrice: Double,
     val availableQuantity: Int
 )
 
 data class ProductDetailResponse(
-    val productId: Int,
+    val productId: Long,
     val name: String,
     val description: String,
-    val price: BigDecimal,
+    val price: Double,
     val options: List<ProductOptionResponse>,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime
