+++ src/main/kotlin/kr/hhplus/be/server/repository/product/ProductRepository.kt
@@ -4,7 +4,7 @@ import kr.hhplus.be.server.domain.product.Product
 
 interface ProductRepository {
     fun save(product: Product): Product
-    fun findById(productId: Int): Product?
+    fun findById(productId: Long): Product?
     fun update(product: Product): Product
-    fun delete(productId: Int)
+    fun delete(productId: Long)
 }
