+++ src/main/kotlin/kr/hhplus/be/server/repository/user/BalanceRepository.kt
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.repository.user
 import kr.hhplus.be.server.domain.user.Balance
 
 interface BalanceRepository {
-    fun findByAccountId(accountId: Int): Balance?
+    fun findByAccountId(accountId: Long): Balance?
     fun update(balance: Balance): Balance
     fun save(balance: Balance): Balance
     fun delete(balance: Balance)
